DEV-1097 Fix attachment delete failure issue when clicked twice

+++ src/app/mail/mail-sidebar/compose-mail/compose-mail.component.ts
@@ -633,7 +633,10 @@ export class ComposeMailComponent implements OnInit, AfterViewInit, OnDestroy {
   }
 
   removeAttachment(attachment: Attachment) {
-    this.store.dispatch(new DeleteAttachment(attachment));
+    if (!attachment.isRemoved) {
+      attachment.isRemoved = true;
+      this.store.dispatch(new DeleteAttachment(attachment));
+    }
   }
 
   updateSignature() {
